Logistic Regression with churn dataset
A telecommunications company is concerned about the number of customers leaving their land-line business for cable competitors. They need to understand who is leaving. 
In this notebook, I use Logistic Regression and create a model for a telecommunication company, to predict when its customers will leave for a competitor so that they can take some action to retain the customers.