Decision Trees with drug200 dataset
In this notebook, I use a popular machine learning algorithm, Decision Trees. I utilize a classification algorithm to build a model from the historical data of patients, and their responses to different medications. Then trained decision tree to predict the class of an unknown patient, or to find a proper drug for a new patient